修复图片处理后下方出现白边bug

+++ Metic/Utils/PhotoGetter.m
@@ -208,7 +208,7 @@
     float adjustWidth = 640.0;
     while (flag) {
         if (compressedImage.size.width> adjustWidth) {
-            CGSize imagesize=CGSizeMake(adjustWidth, compressedImage.size.height * adjustWidth/compressedImage.size.width);
+            CGSize imagesize=CGSizeMake((NSInteger)adjustWidth, (NSInteger)(compressedImage.size.height * adjustWidth/compressedImage.size.width));
             compressedImage = [compressedImage imageByScalingToSize:imagesize];
             imageData = UIImageJPEGRepresentation(compressedImage, 1.0);
         }
@@ -268,13 +268,13 @@
     NSData* imageData2 = UIImageJPEGRepresentation(compressedImage2, 1.0);
     
     if (compressedImage1.size.width> 640) {
-        CGSize imagesize=CGSizeMake(640.0, compressedImage1.size.height * 640.0/compressedImage1.size.width);
+        CGSize imagesize=CGSizeMake(640.0, (NSInteger)(compressedImage1.size.height * 640.0/compressedImage1.size.width));
         compressedImage1 = [compressedImage1 imageByScalingToSize:imagesize];
         imageData1 = UIImageJPEGRepresentation(compressedImage1, 1.0);
     }
     
     if (compressedImage2.size.width> 300) {
-        CGSize imagesize=CGSizeMake(300.0, compressedImage2.size.height * 300.0/compressedImage2.size.width);
+        CGSize imagesize=CGSizeMake(300.0, (NSInteger)(compressedImage2.size.height * 300.0/compressedImage2.size.width));
         compressedImage2 = [compressedImage2 imageByScalingToSize:imagesize];
         imageData2 = UIImageJPEGRepresentation(compressedImage2, 1.0);
     }
@@ -345,7 +345,7 @@
     float adjustWidth = 640.0;
     while (flag) {
         if (compressedImage.size.width> adjustWidth) {
-            CGSize imagesize=CGSizeMake(adjustWidth, compressedImage.size.height * adjustWidth/compressedImage.size.width);
+            CGSize imagesize=CGSizeMake((NSInteger)adjustWidth, (NSInteger)(compressedImage.size.height * adjustWidth/compressedImage.size.width));
             compressedImage = [compressedImage imageByScalingToSize:imagesize];
             imageData = UIImageJPEGRepresentation(compressedImage, 1.0);
         }
